requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
drf-spectacular==0.29.0
gunicorn==25.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
whitenoise==6.12.0
